logic changed on emotion-records & SearchBar according to PR
