+++ pages/emotion-records.js
@@ -1,8 +1,9 @@
 import EmotionRecordsList from "@/components/EmotionRecordsList";
 import styled from "styled-components";
 import SearchBar from "@/components/SearchBar";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import Link from "next/link";
+import Fuse from "fuse.js";
 
 const StyledWrapper = styled.div`
   display: flex;
@@ -55,14 +56,43 @@ export default function EmotionRecords({
   onDeleteEmotionEntry,
   toggleHighlight,
 }) {
-  const [shownEntries, SetShownEntries] = useState(emotionEntries);
+  const [searchTerm, setSearchTerm] = useState();
+  const [shownEntries, setShownEntries] = useState(emotionEntries);
   const [isHighlighted, setIsHighlighted] = useState(false);
+
+  useEffect(() => {
+    if (!searchTerm) {
+      setShownEntries(emotionEntries);
+      return;
+    }
+
+    const fuse = new Fuse(emotionEntries, {
+      includeScore: true,
+      threshold: 0.4,
+      keys: [
+        "date",
+        "tensionLevel",
+        "trigger",
+        "intensity",
+        "notes",
+        "category",
+        "emotion",
+        "subemotion",
+      ],
+    });
+
+    const results = fuse.search(searchTerm);
+    const items = results.map((result) => result.item);
+    setShownEntries(items);
+  }, [emotionEntries, searchTerm]);
 
   function handleShowHighlighted() {
     setIsHighlighted(!isHighlighted);
   }
 
-  console.log(shownEntries);
+  function handleSearch(input) {
+    setSearchTerm(input);
+  }
 
   return (
     <StyledWrapper>
@@ -71,10 +101,7 @@ export default function EmotionRecords({
 
         {emotionEntries.length !== 0 && (
           <>
-            <SearchBar
-              SetShownEntries={SetShownEntries}
-              emotionEntries={emotionEntries}
-            />
+            <SearchBar onSearch={handleSearch} />
 
             <StyledHighlightButton onClick={handleShowHighlighted}>
               {isHighlighted ? "Show all Entries" : "Show highlighted Entries"}
